requirements.txt
langchain
fastapi
faiss-cpu
openai
PyPDF2
python-dotenv
python-multipart
uvicorn[standard]
gunicorn
websockets
networkx
youtube-transcript-api
tiktoken
passlib[bcrypt]
python-jose[cryptography]
Jinja2
fastapi_jwt_auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pymongo
fastapi_mail